+++ CHANGELOG.md
@@ -9,6 +9,10 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+### Fixed
+
+- Fixed logic in `Beaker.cluster.utilization()` to check `job.limits` instead of `job.requests` for the source of truth.
+
 ## [v1.26.2](https://github.com/allenai/beaker-py/releases/tag/v1.26.2) - 2024-02-29
 
 ### Added